third_party/blink/web_tests/external/wpt/html/rendering/replaced-elements/attributes-for-embedded-content-and-images/img-aspect-ratio.tentative-expected.txt

This is a testharness.js-based test.
FAIL Image width and height attributes are used to infer aspect-ratio assert_approx_equals: expected 0.8 +/- 0.001 but got 1
Harness: the test ran to completion.
